+++ fluent.runtime/docs/conf.py
@@ -24,9 +24,9 @@ copyright = '2019, Luke Plant'
 author = 'Luke Plant'
 
 # The short X.Y version
-version = '0.1'
+version = '0.3'
 # The full version, including alpha/beta/rc tags
-release = '0.1'
+release = '0.3'
 
 
 # -- General configuration ---------------------------------------------------