+++ fluent.runtime/CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+fluent.runtime next
+-------------------
+
+fluent.runtime 0.3 (October 23, 2019)
+---------------------------------------
+
+* Added ``fluent.runtime.FluentResource`` and
+  ``fluent.runtime.FluentBundle.add_resource``.
+* Removed ``fluent.runtime.FluentBundle.add_messages``.
+* Replaced ``bundle.format()`` with ``bundle.format_pattern(bundle.get_message().value)``.
+* Added ``fluent.runtime.FluentLocalization`` as main entrypoint for applications.
+
 fluent.runtime 0.2 (September 10, 2019)
 ---------------------------------------